my_career_report/requirements.txt
Jinja2
WeasyPrint
matplotlib
PyYAML
pandas
